+++ pyccel/ast/numpy_wrapper.py
@@ -64,17 +64,24 @@ class PyccelPyArrayObject(DataType, metaclass=Singleton):
 
 def get_numpy_max_acceptable_version_file():
     """
-    Get the macro specifying the last acceptable numpy version. If numpy is more
-    recent than this then deprecation warnings are shown.
+    Get the macro specifying the most recent acceptable NumPy version.
 
-    The last acceptable numpy version is 1.19. If the current version is older
-    than this then the last acceptable numpy version is the current version
+    Get the macro specifying the most recent acceptable NumPy version.
+    If NumPy is more recent than this then deprecation warnings are shown.
+
+    The most recent acceptable NumPy version is 1.19. If the current version is older
+    than this then the last acceptable NumPy version is the current version.
+
+    Returns
+    -------
+    str
+        A string containing the code which defines the macro.
     """
     numpy_max_acceptable_version = [1, 19]
     numpy_current_version = [int(v) for v in np.version.version.split('.')[:2]]
     numpy_api_acceptable_version = min(numpy_max_acceptable_version, numpy_current_version)
-    numpy_api_macro = '# define NPY_NO_DEPRECATED_API NPY_{}_{}_API_VERSION\n'.format(
-        *numpy_api_acceptable_version)
+    major, minor = numpy_api_acceptable_version
+    numpy_api_macro = f'# define NPY_NO_DEPRECATED_API NPY_{major}_{minor}_API_VERSION\n'
 
     return '#ifndef NPY_NO_DEPRECATED_API\n'+ \
             numpy_api_macro+\
@@ -85,14 +92,14 @@ PyArray_Check = FunctionDef(name      = 'PyArray_Check',
                             arguments = [FunctionDefArgument(Variable(dtype=PyccelPyObject(), name = 'o'))],
                             results   = [FunctionDefResult(Variable(dtype=NativeBool(), name='b'))])
 
-# numpy array to c ndarray : function definition in pyccel/stdlib/cwrapper/cwrapper_ndarrays.c
+# NumPy array to c ndarray : function definition in pyccel/stdlib/cwrapper/cwrapper_ndarrays.c
 pyarray_to_ndarray = FunctionDef(
                 name      = 'pyarray_to_ndarray',
                 arguments = [FunctionDefArgument(Variable(name = 'a', dtype = PyccelPyObject(), memory_handling = 'alias'))],
                 body      = [],
                 results   = [FunctionDefResult(Variable(name = 'array', dtype = NativeGeneric()))])
 
-# numpy array check elements : function definition in pyccel/stdlib/cwrapper/cwrapper_ndarrays.c
+# NumPy array check elements : function definition in pyccel/stdlib/cwrapper/cwrapper_ndarrays.c
 pyarray_check = FunctionDef(
                 name      = 'pyarray_check',
                 arguments = [