+++ pyccel/ast/utilities.py
@@ -190,19 +190,6 @@ def builtin_import(expr):
     return []
 
 #==============================================================================
-def get_function_from_ast(ast, func_name):
-    node = None
-    for stmt in ast:
-        if isinstance(stmt, FunctionDef) and str(stmt.name) == func_name:
-            node = stmt
-            break
-
-    if node is None:
-        print('> could not find {}'.format(func_name))
-
-    return node
-
-#==============================================================================
 def split_positional_keyword_arguments(*args):
     """ Create a list of positional arguments and a dictionary of keyword arguments
     """
@@ -345,7 +332,7 @@ def insert_index(expr, pos, index_var):
         return expr[index_var]
 
     else:
-        raise NotImplementedError("Expansion not implemented for type : {}".format(type(expr)))
+        raise NotImplementedError(f"Expansion not implemented for type : {type(expr)}")
 
 #==============================================================================