+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All notable changes to this project will be documented in this file.
 
 ### Deprecated
 
+-   \[INTERNALS\] Removed unused and undocumented function `get_function_from_ast`.
+
 ## \[1.11.2\] - 2024-03-05
 
 ### Added